FontAwesome -> Fontello
Fontello (http://fontello.com/) lets us build our own custom font and (more
importantly) only include the icons we want. This drops us from ~100KB for FA's
css and woff2 down to ~5k. That's a win.

This also has the pleasant bonus of getting rid of the CSS3 validation warnings
that came with FA. Now we're down to only one, and IMO it's a false positive.