Hide psyinf logo when vertical space is scarce; except on index
The problem was that the website is really fast, and on smaller mobile (where
the page title wasn't above-the-fold), it was hard for the user to detect (other
than the red underline for the link) that the page had actually loaded already.

So, when we lack screen height, suppress the psyinf logo so the page title is
above-the-fold. Except on the home page, where we keep it for branding purposes.