replace FZJ and HHU logos w/ crushed SVG versions
They're crisper this way, and smaller.

Props to SVGOMG [1] for crushing the SVGs. Even though the HHU SVG logo is
larger than the PNG, it'll compress quite nicely. The only issue is painting
time, which I havn't bothered to measure yet, but might also be faster than
scaling the PNGs, as the logos aren't all that complex.

[1] https://jakearchibald.github.io/svgomg/